+++ lib/util/folly_validators.dart
@@ -7,100 +7,135 @@ class FollyValidators {
   ///
   ///
   ///
-  static String? decimalGTEZero(Decimal decimal) => decimal.doubleValue >= 0
-      ? null
-      : 'O valor deve ser igual ou maior que zero.';
+  static String? decimalGTEZero(
+    Decimal decimal, {
+    String msg = 'O valor deve ser igual ou maior que zero.',
+  }) =>
+      decimal.doubleValue >= 0 ? null : msg;
 
   ///
   ///
   ///
-  static String? decimalGTZero(Decimal decimal) =>
-      decimal.doubleValue > 0 ? null : 'O valor deve ser maior que zero.';
+  static String? decimalGTZero(
+    Decimal decimal, {
+    String msg = 'O valor deve ser maior que zero.',
+  }) =>
+      decimal.doubleValue > 0 ? null : msg;
 
   ///
   ///
   ///
-  static String? decimalLTZero(Decimal decimal) =>
-      decimal.doubleValue < 0 ? null : 'O valor deve ser menor que zero.';
+  static String? decimalLTZero(
+    Decimal decimal, {
+    String msg = 'O valor deve ser menor que zero.',
+  }) =>
+      decimal.doubleValue < 0 ? null : msg;
 
   ///
   ///
   ///
-  static String? decimalLTEZero(Decimal decimal) => decimal.doubleValue <= 0
-      ? null
-      : 'O valor deve ser igual ou menor que zero.';
+  static String? decimalLTEZero(
+    Decimal decimal, {
+    String msg = 'O valor deve ser igual ou menor que zero.',
+  }) =>
+      decimal.doubleValue <= 0 ? null : msg;
 
   ///
   ///
   ///
-  static String? stringNotEmpty(String? string) =>
-      string != null && string.isNotEmpty
-          ? null
-          : 'O campo não pode ser vazio.';
+  static String? stringNotEmpty(
+    String? string, {
+    String msg = 'O campo não pode ser vazio.',
+  }) =>
+      string != null && string.isNotEmpty ? null : msg;
 
   ///
   ///
   ///
-  static String? stringNullNotEmpty(String? string) =>
-      string == null || string.isNotEmpty
-          ? null
-          : 'Informe um valor ou deixe em branco.';
+  static String? stringNullNotEmpty(
+    String? string, {
+    String msg = 'Informe um valor ou deixe em branco.',
+  }) =>
+      string == null || string.isNotEmpty ? null : msg;
 
   ///
   ///
   ///
-  static String? notNull(dynamic value) =>
-      value == null ? 'O campo não pode ser nulo.' : null;
+  static String? notNull(
+    dynamic value, {
+    String msg = 'O campo não pode ser nulo.',
+  }) =>
+      value == null ? msg : null;
 
   ///
   ///
   ///
-  static String? intGTEZero(int? value) =>
-      (value ?? -1) >= 0 ? null : 'O valor deve ser igual ou maior que zero.';
+  static String? intGTEZero(
+    int? value, {
+    String msg = 'O valor deve ser igual ou maior que zero.',
+  }) =>
+      (value ?? -1) >= 0 ? null : msg;
 
   ///
   ///
   ///
-  static String? intGTZero(int? value) =>
-      (value ?? -1) > 0 ? null : 'O valor deve ser maior que zero.';
+  static String? intGTZero(
+    int? value, {
+    String msg = 'O valor deve ser maior que zero.',
+  }) =>
+      (value ?? -1) > 0 ? null : msg;
 
   ///
   ///
   ///
-  static String? intLTZero(int? value) =>
-      (value ?? 1) < 0 ? null : 'O valor deve ser menor que zero.';
+  static String? intLTZero(
+    int? value, {
+    String msg = 'O valor deve ser menor que zero.',
+  }) =>
+      (value ?? 1) < 0 ? null : msg;
 
   ///
   ///
   ///
-  static String? intLTEZero(int? value) =>
-      (value ?? 1) <= 0 ? null : 'O valor deve ser igual ou menor que zero.';
+  static String? intLTEZero(
+    int? value, {
+    String msg = 'O valor deve ser igual ou menor que zero.',
+  }) =>
+      (value ?? 1) <= 0 ? null : msg;
 
   ///
   ///
   ///
-  static String? intNullGTEZero(int? value) => (value == null || value >= 0)
-      ? null
-      : 'O valor deve ser nulo, igual ou maior que zero.';
+  static String? intNullGTEZero(
+    int? value, {
+    String msg = 'O valor deve ser nulo, igual ou maior que zero.',
+  }) =>
+      (value == null || value >= 0) ? null : msg;
 
   ///
   ///
   ///
-  static String? intNullGTZero(int? value) => (value == null || value > 0)
-      ? null
-      : 'O valor deve ser nulo ou maior que zero.';
+  static String? intNullGTZero(
+    int? value, {
+    String msg = 'O valor deve ser nulo ou maior que zero.',
+  }) =>
+      (value == null || value > 0) ? null : msg;
 
   ///
   ///
   ///
-  static String? intNullLTZero(int? value) => (value == null || value < 0)
-      ? null
-      : 'O valor deve ser nulo ou menor que zero.';
+  static String? intNullLTZero(
+    int? value, {
+    String msg = 'O valor deve ser nulo ou menor que zero.',
+  }) =>
+      (value == null || value < 0) ? null : msg;
 
   ///
   ///
   ///
-  static String? intNullLTEZero(int? value) => (value == null || value <= 0)
-      ? null
-      : 'O valor deve ser nulo, igual ou menor que zero.';
+  static String? intNullLTEZero(
+    int? value, {
+    String msg = 'O valor deve ser nulo, igual ou menor que zero.',
+  }) =>
+      (value == null || value <= 0) ? null : msg;
 }