+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.16.1] - 2022-11-01
+
+* Updating README.md.
+* Adding new tests for FollyValidators.
+* Deprecating AbstractEnumParser.
+
 ## [0.16.0] - 2022-10-30
 
 * Initial pub.dev publish